[26] 장바구니 관련 수정
- PUT mapping URI 설계 변경
- init 메서드와 addItem 메서드 통합
- 동시성 이슈 관련 수정

+++ src/main/java/me/jjeda/mall/cart/controller/CartController.java
@@ -3,6 +3,7 @@ package me.jjeda.mall.cart.controller;
 import lombok.RequiredArgsConstructor;
 import me.jjeda.mall.accounts.dto.AccountDto;
 import me.jjeda.mall.cart.domain.CartItem;
+import me.jjeda.mall.cart.domain.CartModifyType;
 import me.jjeda.mall.cart.service.CartService;
 import me.jjeda.mall.common.CurrentUser;
 import org.springframework.http.ResponseEntity;
@@ -13,6 +14,9 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RequestParam;
+
+import java.util.Objects;
 
 @Controller
 @RequiredArgsConstructor
@@ -20,26 +24,21 @@ import org.springframework.web.bind.annotation.RequestMapping;
 public class CartController {
     private final CartService cartService;
 
-    @PostMapping
-    public ResponseEntity initCart(@CurrentUser AccountDto accountDto) {
-        return ResponseEntity.ok(cartService.initCart(String.valueOf(accountDto.getEmail())));
-    }
-
     @GetMapping
     public ResponseEntity getCart(@CurrentUser AccountDto accountDto) {
         return ResponseEntity.ok(cartService.getCart(String.valueOf(accountDto.getEmail())));
     }
 
-    @PutMapping("/add")
-    public ResponseEntity addItem(@CurrentUser AccountDto accountDto, @RequestBody CartItem cartItem) {
-        return ResponseEntity.ok(cartService.addItem(String.valueOf(accountDto.getEmail()), cartItem));
-    }
+    @PutMapping
+    public ResponseEntity modifyCart(@CurrentUser AccountDto accountDto, @RequestBody CartItem cartItem,
+                                  @RequestParam CartModifyType cartModifyType) {
 
-    @PutMapping("/remove")
-    public ResponseEntity removeItem(@CurrentUser AccountDto accountDto, @RequestBody CartItem cartItem) {
-        return ResponseEntity.ok(cartService.removeItem(String.valueOf(accountDto.getEmail()), cartItem));
+        if(Objects.equals(cartModifyType,CartModifyType.ADD)) {
+            return ResponseEntity.ok(cartService.addItem(String.valueOf(accountDto.getEmail()), cartItem));
+        } else {
+            return ResponseEntity.ok(cartService.removeItem(String.valueOf(accountDto.getEmail()), cartItem));
+        }
     }
-
     @DeleteMapping
     public ResponseEntity deleteCart(@CurrentUser AccountDto accountDto) {
         cartService.deleteCart(String.valueOf(accountDto.getId()));
